requirements.txt
streamlit
langgraph
langchain
faiss-cpu
sentence-transformers
pymupdf
transformers
torch